requirements.txt
Flask==3.1.1
pyyaml==5.1
